Remove status-interval
There is no reason to set this.

This plugin still does not need tmux to poll for updates (we still use
interrupts, and setting status-interval = 0 will keep working), but on
the other hand there is no reason to disable the update interval
completely just because I think that it is a saner default.

If you don't need tmux statusline polling, I still recommend disabling
it by using set -g status-interval 0, but that is a general tmux
recommendation and has nothing to do with this plugin, so we should
remove it from the documentation.