Add support for the new ModeChanged autocmd
Look at 224e79d to learn about why we preemptively use deferred_update().
The tl;dr is that mappings that use ":" would very briefly show Command
mode, which appears as a distracting flickering in the statusline.
We defer the update for a few milliseconds to fix the problem.

In the future we should probably find a better solution that updates the
statusline immediately.

Fixes #11