Improve latency when entering visual mode
Only defer the update when we really need to.
See bf15307 and 224e79d for why we defer the update at all.
However, this workaround is only needed for CmdlineEnter to fix the
problem where executed mappings could trigger a statusline update which
would appear as a flickering statusline.

As a solution we update immediately on ModeChanged in general except for
CmdlineEnter, which we handle separately in an extra autocmd.