Use the new nvim_eval_statusline API
Using this brand new neovim-only API has a few advantages:

- It is faster than our current vimscript implementation
- It can evaluate the statusline within a specific winid context, which
  fixes some weird problems with statuslines that check if the current
  evaluater is the currently active buffer. For example lualine checks
  this using vim.g.statusline_winid == vim.fn.win_getid(), which does
  not work at all with the previous method.
- It should be more reliable in general.

This API requires neovim 0.6 or a nightly neovim compiled after
2021-10-18, i.e. one including upstream neovim commit
9086938f7bd6d6ccb7f4a30fb78aeaf0d84e4471.

A small problem to consider for the future:
Right now nvim_eval_statusline does not support returning an index,
where the left part of the statusline ends and where the right part of
the statusline starts. We fix this by using a weird workaround with the
fillchar argument, but this may break if the terminal window gets too
small, so we should probably exchange it with something more reliable in
the future.

Right now we are using the ASCII Unit Separator (031, type <C-V>031 in
insert mode) as fillchar, this should be fine for the time being.

Also we should probably make use of the use_tabline option, if
g:tpipeline_tabline is set.

Fixes #13