Disable ruler by default
If ruler is set (which is set by default), it will only show the ruler
if there is no statusline set OR if laststatus=0.
But since this plugin sets laststatus=0, the ruler will suddenly appear
inside tmux even though it does not appear outside of tmux (since
laststatus != 0).

We can assume that people that use this plugin use a statusline.
Therefore the overall default behaviour for users of this plugin is that
ruler is not shown, because stl is non-empty (by assumption) and
laststatus!=0 (by assumption).
Therefore we should use 'set noruler' to match that behaviour inside
tmux. Otherwise ruler will be shown, because laststatus=0.

Usually we should not change unrelated basic vim settings from their
default value, but this is an exemption to the rule, because we don't
change the default behaviour.

Fixes #15