Implement support for text intensity
We initially support bold and italic.

Note that tmux remembers the text intensity. This means that if in tmux
a block wants bold with [fg=#ffffff,bg=#000000,bold] and then another
block follows that does not want bold, e.g. [fg=#ffffff,bg=#000000] then
tmux will still make the text bold.
We need to explicitly use [fg=#ffffff,bg=#000000,nobold] to disable bold
for the next block.
Therefore we store Booleans (s:b and s:i) for whether we need to restore
nobold or noitalics in the next block.

We still need to see if this works out for the vim statusline ending
with a bold text (and therefore not resetting bold for tmux). If the
user has other content after this in the tmux statusline, it may
accidentally be rendered in bold. For now we leave it as is and maybe we
need to add a workaround in the future.

This implements one part for #16, the other part is respecting the color
in the middle of the tmux statusline.