Allow to fill the space in the statusline centre
The default vim statusline colors the middle of the statusline with the
last color used in the left part. This was previously not used in this
plugin, we instead just use the default background color in order to not
interfere with any stuff in the middle of the tmux statusline.

However users should still be able to also use this feature in the tmux
statusline, so they can opt-in to this now with:
let g:tpipeline_fillcentre = 1

This implements the final part to fix #16

You could try it out for example by using lualine:

lua require'lualine'.setup{options = {theme = 'powerline'}}
let g:tpipeline_statusline=&stl