Update README
I can't reproduce this anymore on latest neovim so let's remove the
notice.

In the future we should think about making g:tpipeline_cursormoved = 1
the default.
However, let's push this back for now until we have done the planned Lua
Rewrite of the Core Update function for neovim, which will probably fix
that problem for old neovim instances too due to the performance
advantage over vimscript8.