Add workaround to hide statusline for split windows
Enable this feature with let g:tpipeline_clearstl = 1.
This will clear the vim internal statusline after the plugin has been
loaded, so that the statusline does not appear as duplicate statusline
for split windows.

In the future we could add an option to customize the appearance of the
cleared statusline, but for now '%#StatusLine#' should be a good
default.

Fixes #19