Cleanup properly
Previously there was a race condition where if the running job was slow,
it could write into the socket right after the writefile() in cleanup()
completed, causing the statusline to remain even after vim quit.

Now we explicitly kill the job to prevent this from happening.

To reproduce this 100% reliably before this patch, you can add a sleep 1
in the dispatched job before writing to the socket and a sleep 3 at the
end of the cleanup() method.