Move FocusLost init outside of condition
The cautious_cleanup heuristic works best, if the readfile() call
happens as late as possible. This is because this way we detect vim
switching to another vim pane in the most reliable way. If we wait too
short, we do not see that another instance will write to the socket in
the next millisecond(s) and will do an unsolicited update.
Therefore it is better to defer the readfile() to the last possible
moment, so that we detect this scenario in more cases.