Use the entire terminal width in nvim_eval_statusline()
I paid for the entire terminal width, I'm gonna use the entire terminal.

nvim_eval_statusline() uses the current vim window size, if no maxwidth
parameter is given. We don't want that, as we can use the size of the
entire terminal, which can be many times larger than the size of the
current vim window.

For example consider the following:
- Open tmux
- Split into 2 tmux panes
- Open vim in one pane and decrease the width of that pane

Expected behaviour:
Vim statusline appears in the entire tmux statusline.

Actual behaviour:
Vim statusline appears only in the left part of the tmux statusline and
is cropped to the width of the vim window.

This is triggered only because we use g:tpipeline_fillchar to split the
statusline into two parts, but there is no space to fill with fillchar
because we run out of space. The clean solution would be to have neovim
somehow return the location of the split, but it doesn't really matter
that much because this behaviour normally only occurs only (apart from
above scenario) when the statusline spans the entire tmux statusline, in
which case it doesn't matter where we split because it fills the entire
bottom anyway.

For the above scenario we now query the terminal width and pass that to
nvim_eval_statusline(). Right now this is done with a blocking call on
startup, we should probably do that in an async kind of way. For now
this is okay though, we will just tackle this when we rewrite that part
in Lua anyway.

Fixes #25