Translate uppercase color codes
synIDtrans() returns color codes in the case that they are defined with,
i.e. it is case-sensitive. Vim does not have a problem with uppercase
color codes, but tmux does have one.

Therefore we have to transpile the color codes from uppercase to
lowercase, so that tmux stops whining.

An example theme that uses uppercase color codes, is catppuccin [1]

Fixes #26

[1] https://github.com/catppuccin/nvim/blob/773d339cbd307fe218cf7b3ea04eac26b345a3b7/lua/catppuccin/core/palettes/latte.lua#L10