Revert "Recommend absolute-centre directly"
This reverts commit 3b110dd00850b22a8e6fdc87d3ec887cac40e6d9.

People are still using very old tmux versions like 3.0, see #27.

Also maybe absolute-centre is not the best default in the first place,
since it doesn't deal as well with limited space issues. The only reason
for recommending absolute-centre was that it stops the flickering when
switching tmux windows and the next window had a different ratio of
left-right statusline usage.
For now let's default back to centre instead, and wait a bit more.