Improve width calculation
We have to ignore possible shell output previous to the real tmux
output.
For example the output could be:

Agent pid 17835
238

We do this with a simple heuristic of just caring about the last line.
Hopefully this works for all cases.
And hopefully there won't be a case in the future where some junk output
comes **after** the real tmux output.

What a total shitshow...

Fixes #27