Do not load in GUI vim by default
Obviously if GUI vim is started outside of tmux, we always already
disabled this plugin, since $TMUX would be empty then.
But if someone starts GUI vim from the command line within tmux, then
$TMUX would not be empty and the plugin would become enabled which would
look very weird, because we then embed the GUI vim statusline in a
remote tmux console window.

For gVim fixing this is very easy, as we can simply check for
has('gui_running').

Unfortunately, this variable will probably not work in neovim GUIs,
since the neovim GUI only attaches after vimrc has already been sourced,
so has('gui_running') will return 0 even though the GUI will attach
later in the startup process.

We do not want to use neovim-qt specific variables like g:GuiLoaded [1].
For example gonvim uses another variable [2].

This is where things get more complicated.
It seems like the only reliable way to detect it is to use the UIEnter
autocmd [3], which luckily exists since 2017 [4].

There is also a way to identify neovim clients over RPC [5], but using
autocmds is probably much easier.

Therefore to solve the problem, we disable the plugin on UIEnter, which
we do with the new tpipeline#state#restore() function added in
a95ccb6939c819142f9b972033010403c4b7f5d8.

Theoretically we could also reload the plugin again on UILeave, but in
reality I don't see anyone ever needing to do this.

We could also add an option to keep the plugin enabled for GUIs if
someone wants to have that feature, but let's do that in the future if
that usecase ever comes up, which it probably won't (yes I am jinxing
it).
At least we now have sane defaults for GUI vim.

Fixes #29.

[1] https://github.com/equalsraf/neovim-qt/issues/219
[2] https://github.com/dzhou121/gonvim/issues/17
[3] https://github.com/neovim/neovim/issues/3646
[4] https://github.com/neovim/neovim/pull/6917
[5] https://github.com/neovim/neovim/pull/6743