Postpone evaluation of stl for nvim
Usually the statusline is set once at the startup of vim to a function
that evaluates to the dynamic value of the statusline at runtime.

However some statuslines are particularly retarded and completely shit
on this vim idiomatic way of setting the statusline and instead update
the stl vim option directly, e.g. with a timer and on autocmd events.

This is not only completely stupid, but also has slightly worse
performance and is in fact also semantically incorrect (e.g. there is
not an autocmd for every way that it is possible to update the
statusline, only for most of the scenarios).

This is now the second time that lualine does something in a completely
fucked up way (the first time was the way that they pretend to have
global statusline support, by simply setting the laststatus option if a
lualine option is set, completely ignoring the idiomatic way of simply
letting the user set the laststatus option directly).
Why again are so many people recommending lualine?
They seem to know jack shit about the idiomatic way to do things in vim.

The bad commit in lualine is 53aa3d82d9d9329880b4197f278fcb74051a6af5.

Of course this also breaks our plugin because of the way that
g:tpipeline_statusline interacts with &stl.

To workaround the issue, we do not set g:tpipeline_statusline if we are
in neovim. Then in the update routine we first evaluate
g:tpipeline_statusline (to not break users configs that set this
variable) and then try &stl as fallback.

For now we only implement this workaround for neovim, as I don't believe
there is any vim statusline plugin that uses stl in such a fucked up
way.

Congrats lualine, we know have an extra workaround just for you.

Fixes #32

ref: https://github.com/nvim-lualine/lualine.nvim/commit/53aa3d82d9d9329880b4197f278fcb74051a6af5
ref: https://github.com/nvim-lualine/lualine.nvim/pull/736