Relax assumptions about early stl evaluation
In affe9ca352dabb8eddbcedd0baac4672f9173724 we already postponed the
evaluation for neovim. This was because lualine used some weird
shenanigans to update the statusline that were incompatible with early
evaluation of the statusline.

It looks like lualine is not the only statusline that behaves like that
and it turns out that we can even get the same scenario in vim by using
lightline + vim-lightline-coc.

Thus we need to postpone the evaluation for vim as well.

Fixes #33