Bump required vim version
The new minimum required versions are:
- vim 8.2.4650 OR
- neovim 0.6

Users who can't satisfy these dependencies, should pin the version of
this plugin to v1.0

We have to raise the minimum vim requirement to vim 8.2.4650, because we
now use vim9script for better performance. The speedup is worth the
tradeoff of dropping support for older vim versions.

This also simplifies a bit of code.

Since our neovim code uses nvim_eval_statusline(), compatibility of
neovim is not affected. However since we drop the legacy vimscript path,
we now require at least neovim 0.6, since that is the first release with
nvim_eval_statusline().
Note that we still recommend to use neovim 0.7, but neovim 0.6 is still
supported as well.

Vim users who use versions older than 8.2.4650, should switch to neovim
or alternatively pin this plugin to v1.0