Fix type error in Parse_stl
eval() can sometimes return a Number.
For example if we evaluate an inner function that returns a number, we
will then pass the wrong type to Parse_stl.

Fix this by always converting to a string by concatenating with an empty
string (which converts simple types to string automatically).

We could probably also fix this by using string() or by using a
printf-style format string.