Subscribe to stl OptionSet for g:tpipeline_clearstl
Before this patch, after we painstakingly cleared the split statusline,
all would be ruined again if stl is set to something different again.

To make matters worse we would not even adapt the main statusline to the
newly set stl, because we only listen to g:tpipeline_statusline changes.

Fix this by subscribing to stl updates and then updating
g:tpipeline_statusline again and clearing stl.

Only enable this if g:tpipeline_clearstl is set and even then I am not
really sure if this is really the correct default thing to do.

Right now we are on about the 11th layer of workarounds to have this
working without issues, so we need to look out in the future if we want
to revert something.

At this point it might be best to go for an upstream vim patch that
allows us to set the split line separately from the main statusline,
so that we can clear it without all of these workarounds.

See #36