Remove remnants of delayed_update()
This workaround wasn't in use anymore since 776026d7311001495095f0b49dd2f0ebf33c3604

Let's remove the last references to it. If regressions start to appear,
we can still revert this.

This also removes the job_check, but this was hardly necessary and
something similar to 8b508f763dc6e8df0d3cac2721f51ebe215b6228 could be
added with a separate timer.