Only adopt stl option updates in global scope
Otherwise if some window updates the stl option locally and doesn't
reset if afterwards, it may be stuck there forever.
For example this was the case with CocList outline.

This may be improved on further in the future by storing a list of local
stl option values and selectively switching to them if that window is
the currently active window.

For now just ignoring these local stl option updates should be fine
though.