Fix initialization order
The initialization order was wrong, it was possible for tpipeline to
try to setup the hooks without fundamental variables being set.

This can happen if v:vim_did_enter is true (which happens e.g. if the
plugin is currently not installed, then installed via vim-plug, and then
activated during runtime).
In that case init_statusline() would be called without s:has_modechgd
being initialized, which would show an ugly warning to the user.

There is no real problem to be found here and the error would not appear
again when restarting vim, but still it is not a good first impression
when the user sees such an unnecessary error.

To fix this, delay init_statusline() until all necessary variables are
set, even if v:vim_did_enter is true.

Fixes #47