Allow to strip highlights in tests
This makes the tests more robust, as we can now use assert_equal()
instead of the weaker assert_match().

We need to strip highlights, because neovim by default always prefixes
the statusline with the 'StatusLine' highlight group in
nvim_eval_statusline(), which is not the case for our vim evaluation.