Disable updates after VimLeavePre
Otherwise we may accidentally trigger a chansend() when the
corresponding job is already dead, which will print an error message
like:

Error detected while processing function <lambda>2[1]..tpipeline#lualine#fix_stl[9]..tpipeline#update:
line   34:
Can't send data to closed stream

This issue only happened in the lualine codepath (since this is the only
part where we are using timers).
It was already fixed upstream with neovim switching the TUI to a
separate process [0].
However, let's fix this on this plugin site for neovim 0.8 users.

Also in neovim 0.8 debug builds there is an issue with neovim taking 2
seconds to exit with lualine. This is not related to this plugin, but
lualine's fault for not closing all libuv handles [1]. Get your shit
together, lualine!

Fixes #56

[0] https://github.com/neovim/neovim/pull/18375
[1] https://github.com/neovim/neovim/issues/18670