Check that colorscheme has true color support
Unfortunately neovim has no support for hlget() yet [0], so this got a
little bit more complicated there.

Also what the hell is wrong with Lua to overcomplicate basic things such
as getting the name of the colorscheme. Apparently we have to use the
generic vimscript command interface [1] just for this.

It's kind of funny how the Lua API now has `api.nvim_cmd()`,
`api.nvim_exec()` and `api.nvim_exec2()` similar to how libc has evolved
to have 1000 alternatives to strcpy() and every year they release a new
version that is apparently safer.
So when will we get `api.nvim_exec3()`? ;)

[0] https://github.com/neovim/neovim/pull/16270
[1] https://github.com/neovim/neovim/issues/18201