Ignore fcs chars in the default statusline
Otherwise we get the weird "-" fillchars outside of tmux in the middle
of our statusline. We really only want them to appear when they are part
of the border between two windows.
Unfortunately vim does not allow to configure the fillchars for the real
statusline and for border lines separately (most people also don't have
border lines, because they don't have something like laststatus=0 plus
fcs set, so for them the real statusline renders inbetween windows).

To fix this, we use the Ignore highlight group to color the middle
section, which just happens to make the fillchars transparent.
This statusline is not evaluated for window borders, so luckily they
still appear there with the original highlight group.