Avoid errors in search call
If the user uses an invalid search query (for example "\(", which has an
unmatched '('), then the searchcount() call will throw an error.
Since the statusline is evaluated regularly, this would then spam the
cmdline messages with the same error over and over again (E54 in the
above case).

We can safely suppress errors here, because the user will still get
notified of invalid queries when doing the initial search. In the error
case searchcount() will return a result that represents "0/0" search
results, so we don't need to do any further error handling