kitty: Do not copy linefeed
The statusline pipe **always** contains a LF character at the end. This
is needed, because the "#(cat ...)" call for the tmux statusline works
in a line-oriented way and actually requires the LF to make the
statusline appear.

Up until recently the kitty statusline worked just fine with that LF,
but it seems that now kitty displays a blank line if we add a LF to the
statusline (which is not entirely unreasonably by kitty).

Therefore we remove the LF from the read pipe, which we can just
hardcode, because it will always be there.

Fixes #65