Split up stdin piping
This splits up the part where we read stdin and where we write it into
the pipe.

It's a shame that this breaks the consistency of beginning each string
append operation with "; ", but we can blame shell "do" syntax for this.

The intention of this patch is to make a follow-up patch more clear,
that implements batching of updates for downstream consumers.